# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "studyplan API",
    "description": "学习规划助手 V1.1 —— LangGraph 编排 + FastAPI 接入 + PostgreSQL 业务事实源。\n\n**契约真相源是 Pydantic 模型**，本 OpenAPI 是其导出物（ADR-0004）。\n**AuthContext 永不出现在可写请求体**，由服务端会话派生。",
    "version": "1.1.0"
  },
  "paths": {
    "/api/v1/resources/searches": {
      "post": {
        "summary": "Search Resources",
        "operationId": "search_learning_resources",
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "description": "学习空间 ID；必须属于当前会话的项目范围，否则 403",
              "title": "Project Id"
            },
            "description": "学习空间 ID；必须属于当前会话的项目范围，否则 403"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResourceSearchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceSearchView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/resources/inspections": {
      "post": {
        "summary": "Inspect Resource",
        "operationId": "inspect_learning_resource",
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "description": "学习空间 ID；必须属于当前会话的项目范围，否则 403",
              "title": "Project Id"
            },
            "description": "学习空间 ID；必须属于当前会话的项目范围，否则 403"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResourceInspectionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceInspectionView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/resources/inspections/{inspection_id}": {
      "get": {
        "summary": "Get Inspection",
        "operationId": "get_resource_inspection",
        "parameters": [
          {
            "name": "inspection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Inspection Id"
            }
          },
          {
            "name": "project_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "description": "学习空间 ID；必须属于当前会话的项目范围，否则 403",
              "title": "Project Id"
            },
            "description": "学习空间 ID；必须属于当前会话的项目范围，否则 403"
          },
          {
            "name": "plan_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 512,
              "title": "Plan Id"
            }
          },
          {
            "name": "stage_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 512,
              "title": "Stage Id"
            }
          },
          {
            "name": "unit_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 512,
              "title": "Unit Id"
            }
          },
          {
            "name": "node_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 512
                },
                {
                  "type": "null"
                }
              ],
              "title": "Node Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceInspectionView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/resources/inspections/by-key/{key}": {
      "get": {
        "summary": "Get Inspection By Key",
        "operationId": "get_resource_inspection_by_key",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          },
          {
            "name": "project_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "description": "学习空间 ID；必须属于当前会话的项目范围，否则 403",
              "title": "Project Id"
            },
            "description": "学习空间 ID；必须属于当前会话的项目范围，否则 403"
          },
          {
            "name": "plan_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 512,
              "title": "Plan Id"
            }
          },
          {
            "name": "stage_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 512,
              "title": "Stage Id"
            }
          },
          {
            "name": "unit_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 512,
              "title": "Unit Id"
            }
          },
          {
            "name": "node_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 512
                },
                {
                  "type": "null"
                }
              ],
              "title": "Node Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceInspectionView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/resources/searches/{search_id}": {
      "get": {
        "summary": "Get Search",
        "operationId": "get_resource_search",
        "parameters": [
          {
            "name": "search_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Search Id"
            }
          },
          {
            "name": "project_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "description": "学习空间 ID；必须属于当前会话的项目范围，否则 403",
              "title": "Project Id"
            },
            "description": "学习空间 ID；必须属于当前会话的项目范围，否则 403"
          },
          {
            "name": "plan_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 512,
              "title": "Plan Id"
            }
          },
          {
            "name": "stage_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 512,
              "title": "Stage Id"
            }
          },
          {
            "name": "unit_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 512,
              "title": "Unit Id"
            }
          },
          {
            "name": "node_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 512
                },
                {
                  "type": "null"
                }
              ],
              "title": "Node Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceSearchView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/resources/searches/by-key/{key}": {
      "get": {
        "summary": "Get Search By Key",
        "operationId": "get_resource_search_by_key",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          },
          {
            "name": "project_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "description": "学习空间 ID；必须属于当前会话的项目范围，否则 403",
              "title": "Project Id"
            },
            "description": "学习空间 ID；必须属于当前会话的项目范围，否则 403"
          },
          {
            "name": "plan_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 512,
              "title": "Plan Id"
            }
          },
          {
            "name": "stage_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 512,
              "title": "Stage Id"
            }
          },
          {
            "name": "unit_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 512,
              "title": "Unit Id"
            }
          },
          {
            "name": "node_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 512
                },
                {
                  "type": "null"
                }
              ],
              "title": "Node Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceSearchView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/resources/selections": {
      "get": {
        "summary": "List Selected",
        "operationId": "list_selected_resources",
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "description": "学习空间 ID；必须属于当前会话的项目范围，否则 403",
              "title": "Project Id"
            },
            "description": "学习空间 ID；必须属于当前会话的项目范围，否则 403"
          },
          {
            "name": "plan_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 512,
              "title": "Plan Id"
            }
          },
          {
            "name": "stage_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 512,
              "title": "Stage Id"
            }
          },
          {
            "name": "unit_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 512,
              "title": "Unit Id"
            }
          },
          {
            "name": "node_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 512
                },
                {
                  "type": "null"
                }
              ],
              "title": "Node Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SelectedResourceView"
                  },
                  "title": "Response List Selected Resources"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Select Resource",
        "operationId": "select_learning_resource",
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "description": "学习空间 ID；必须属于当前会话的项目范围，否则 403",
              "title": "Project Id"
            },
            "description": "学习空间 ID；必须属于当前会话的项目范围，否则 403"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResourceSelectionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelectedResourceView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/resources/manual": {
      "post": {
        "summary": "Manual Resource",
        "operationId": "add_manual_learning_resource",
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "description": "学习空间 ID；必须属于当前会话的项目范围，否则 403",
              "title": "Project Id"
            },
            "description": "学习空间 ID；必须属于当前会话的项目范围，否则 403"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualResourceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelectedResourceView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/resources/selections/{selection_id}": {
      "delete": {
        "summary": "Remove Resource",
        "operationId": "remove_learning_resource",
        "parameters": [
          {
            "name": "selection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Selection Id"
            }
          },
          {
            "name": "project_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "description": "学习空间 ID；必须属于当前会话的项目范围，否则 403",
              "title": "Project Id"
            },
            "description": "学习空间 ID；必须属于当前会话的项目范围，否则 403"
          },
          {
            "name": "plan_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 512,
              "title": "Plan Id"
            }
          },
          {
            "name": "stage_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 512,
              "title": "Stage Id"
            }
          },
          {
            "name": "unit_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 512,
              "title": "Unit Id"
            }
          },
          {
            "name": "node_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 1,
                  "maxLength": 512
                },
                {
                  "type": "null"
                }
              ],
              "title": "Node Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceRemovalView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JsonValue": {},
...
      "ResourceInspectionRequest": {
        "properties": {
          "plan_id": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Plan Id"
          },
          "stage_id": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Stage Id"
          },
          "unit_id": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Unit Id"
          },
          "node_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 512,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Node Id"
          },
          "search_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Search Id"
          },
          "candidate_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Candidate Id"
          },
          "idempotency_key": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Idempotency Key"
          },
          "paths": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 2,
            "title": "Paths"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "plan_id",
          "stage_id",
          "unit_id",
          "search_id",
          "candidate_id",
          "idempotency_key"
        ],
        "title": "ResourceInspectionRequest"
      },
      "ResourceInspectionView": {
        "properties": {
          "inspection_id": {
            "type": "string",
            "title": "Inspection Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "dispatched",
              "succeeded",
              "failed",
              "reconciliation_required"
            ],
            "title": "Status"
          },
          "search_id": {
            "type": "string",
            "title": "Search Id"
          },
          "candidate_id": {
            "type": "string",
            "title": "Candidate Id"
          },
          "candidate": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResourceCandidateView"
              },
              {
                "type": "null"
              }
            ]
          },
          "receipts": {
            "items": {
              "additionalProperties": {
                "$ref": "#/components/schemas/JsonValue"
              },
              "type": "object"
            },
            "type": "array",
            "title": "Receipts"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "inspection_id",
          "status",
          "search_id",
          "candidate_id",
          "created_at"
        ],
        "title": "ResourceInspectionView"
      },
...
      "ResourceRemovalView": {
        "properties": {
          "removed": {
            "type": "boolean",
            "title": "Removed"
          }
        },
        "type": "object",
        "required": [
          "removed"
        ],
        "title": "ResourceRemovalView"
      },
      "ResourceSearchRequest": {
        "properties": {
          "plan_id": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Plan Id"
          },
          "stage_id": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Stage Id"
          },
          "unit_id": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "title": "Unit Id"
          },
          "node_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 512,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Node Id"
          },
          "source": {
            "type": "string",
            "enum": [
              "web",
              "github"
            ],
            "title": "Source",
            "default": "web"
          },
          "query": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Query"
          },
          "idempotency_key": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Idempotency Key"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "plan_id",
          "stage_id",
          "unit_id",
          "query",
          "idempotency_key"
        ],
        "title": "ResourceSearchRequest"
      },
      "ResourceSearchView": {
        "properties": {
          "search_id": {
            "type": "string",
            "title": "Search Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "dispatched",
              "succeeded",
              "failed",
              "reconciliation_required"
            ],
            "title": "Status"
          },
          "query": {
            "type": "string",
            "title": "Query"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/ResourceCandidateView"
            },
            "type": "array",
            "title": "Candidates"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "source": {
            "type": "string",
            "enum": [
              "web",
              "github"
            ],
            "title": "Source",
            "default": "web"
          },
          "context_snapshot": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "type": "object",
            "title": "Context Snapshot"
          }
        },
        "type": "object",
        "required": [
          "search_id",
          "status",
          "query",
          "candidates",
          "created_at"
        ],
        "title": "ResourceSearchView"
      },
...
    }
  }
}